+++ test/device/atomics.jl
