+++ test/device/core.jl
