+++ test/device/hints.jl
