+++ test/device/math.jl
