+++ test/device/reductions.jl
